+++ R/are_equations.R
@@ -1,0 +1,17 @@
+is_eq <- function(x) {
+    inherits(x, "eq")
+}
+
+is_eqs <- function(x) {
+    inherits(x, "eqs")
+}
+
+are_equations <- function(x) {
+    if (rlang::is_null(x)) {
+        return(FALSE)
+    }
+
+    purrr::map_lgl(x, is_eq)
+}
+
+