+++ tests/testthat/test-get_eq_names.R
@@ -1,4 +1,4 @@
-context("test-covariates_name")
+context("test-get_eq_names")
 
 eq_test <- eq(age = 0.1, bmi = -0.3,
     name    = "eq_test",
@@ -14,9 +14,5 @@ eq2_test <- eq(age = 0.1, bmi = -0.3,
 eqs_test <- eqs(eq_test, eq2_test, name = "eqs-test")
 
 test_that("works for eq objects", {
-  expect_equal(get_covariates(eq_test), c("age", "bmi"))
-})
-
-test_that("works for eqs objects", {
-  expect_equal(get_covariates(eqs_test), c("age", "bmi"))
+  expect_equal(get_eq_names(eqs_test), c("eq_test", "eq2_test"))
 })