+++ NEWS.md
@@ -1,3 +1,9 @@
+* Added `eqs` class including costructor, extractors and main methods,
+  to collect equations from the same "reference" and evaluate them
+  together.
+* Changed strata atributes to list of factors
+* Added support for tidy evaluation (`rlang` pkg main utilities)
+
 # equationer 0.0.0.9001
 
 * Added basic methods for class `eq`;